Quarantine Algorithm Added
Changes may need to be made given the placement of the core.py and QAlgo.py files. If this is the case, the code should be readable enough in order to modify the paths without modifying the functions. If additional assistance is required, feel free to contact me; I should be able to answer any questions about it.